+++ src/core/Fieldset/index.ts
@@ -1,0 +1,7 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+export { Fieldset } from './Fieldset';
+export type { FieldsetProps } from './Fieldset';
+export default './Fieldset';