+++ src/core/index.ts
@@ -48,6 +48,9 @@ export type {
 export { ExpandableBlock } from './ExpandableBlock';
 export type { ExpandableBlockProps } from './ExpandableBlock';
 
+export { Fieldset } from './Fieldset';
+export type { FieldsetProps } from './Fieldset';
+
 export { FileUpload, FileUploadTemplate } from './FileUpload';
 export type { FileUploadProps, FileUploadTemplateProps } from './FileUpload';